fix:player in door

+++ include/player.h
@@ -6,7 +6,7 @@
 /*   By: cblonde <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/09/17 12:31:15 by cblonde           #+#    #+#             */
-/*   Updated: 2024/11/21 15:47:28 by cblonde          ###   ########.fr       */
+/*   Updated: 2024/11/25 09:01:52 by cblonde          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -69,7 +69,7 @@ typedef struct s_player
 }	t_player;
 
 t_dir	get_direction(char c);
-int		inside_a_wall(t_data *data, int x, int y, t_map *m);
+int		inside_a_wall(t_data *data, double x, double y, t_map *m);
 void	move_player(t_data *data, int keysym, double speed);
 void	rotate_player(t_data *data, double speed);
 int		handle_input(t_data *data);
